backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.18
requests==2.32.3
dnspython==2.7.0
python-whois==0.9.5
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.28.1
speedtest-cli==2.1.3
ping3==4.0.8
python-dateutil==2.9.0
aiohttp==3.10.11
geopy==2.4.1
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
slowapi==0.1.9

